worker/requirements.txt
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.7
httpx==0.28.1
pdfplumber==0.9.0
# opencv-python==4.13.0.92  # Добавлен обратно если понадобится таблицы
# camelot-py[base]==0.10.1  # отключен из-за OpenCV зависимостей
imaplib2==3.6
requests==2.31.0